+++ services/scanners/ScannerFile.php
@@ -269,6 +269,14 @@ abstract class ScannerFile extends \yii\console\controllers\MessageController
      */
     protected function isValidCategory($category)
     {
-        return !in_array($category, $this->module->ignoredCategories);
+        if ($this->module->onlyCategories) {
+            return in_array($category, $this->module->onlyCategories);
+        }
+
+        if ($this->module->ignoredCategories) {
+            return !in_array($category, $this->module->ignoredCategories);
+        }
+
+        return true;
     }
 }